Active workspace reducer almost ready. Needs testing
