+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/actions/workspaces/activeWorkspace.ts
@@ -19,7 +19,15 @@ import promisify from "~/util/promisify";
 import { AnyActionType, SpecificActionType } from "../index";
 import { reuseExistantValue } from "./helpers";
 import { Dispatch } from "react-redux";
-import { SelectItem, UpdateWorkspaceStateType } from "./index";
+import {
+  SelectItem,
+  UpdateWorkspaceStateType,
+  workspaceStudentsQueryDataType,
+} from "./index";
+import {
+  WorkspaceStaffListType,
+  WorkspaceStudentListType,
+} from "~/reducers/user-index";
 
 /**
  * ACTIVE_WORKSPACE_SET_WORKSPACE_DATA
@@ -105,10 +113,121 @@ export interface UpdateActiveWorkspaceAssessmentRequestTrigger {
 }
 
 /**
- * LoadWorkspaceChatStatusTriggerType
+ * LoadActiveWorkspaceChatStatusTrigger
  */
 export interface LoadActiveWorkspaceChatStatusTrigger {
   (): AnyActionType;
+}
+
+/**
+ * LoadUsersOfWorkspaceTrigger
+ */
+export interface LoadUsersOfActiveWorkspaceTrigger {
+  (data: {
+    workspace: WorkspaceType;
+    payload?: {
+      q: string;
+      active?: boolean;
+      firstResult?: number;
+      maxResults?: number;
+    };
+    success?: (
+      students: WorkspaceStudentListType | WorkspaceStaffListType
+    ) => void;
+    fail?: () => void;
+  }): AnyActionType;
+}
+
+/**
+ * UpdateActiveWorkspaceProducersTrigger
+ */
+export interface UpdateActiveWorkspaceProducersTrigger {
+  (data: {
+    appliedProducers: Array<WorkspaceProducerType>;
+    success: () => void;
+    fail: () => void;
+  }): AnyActionType;
+}
+
+/**
+ * UpdateActiveWorkspaceImagesB64Trigger
+ */
+export interface UpdateActiveWorkspaceImagesB64Trigger {
+  (data?: {
+    delete?: boolean;
+    originalB64?: string;
+    croppedB64?: string;
+    success?: () => void;
+    fail?: () => void;
+  }): AnyActionType;
+}
+
+/**
+ * UpdateActiveWorkspaceDetailsTrigger
+ */
+export interface UpdateActiveWorkspaceDetailsTrigger {
+  (data: {
+    newDetails: WorkspaceDetailsType;
+    success: () => void;
+    fail: () => void;
+  }): AnyActionType;
+}
+
+/**
+ * LoadActiveWorkspaceUserGroupPermissionsTrigger
+ */
+export interface LoadActiveWorkspaceUserGroupPermissionsTrigger {
+  (): AnyActionType;
+}
+
+/**
+ * LoadActiveWorkspaceDetailsTrigger
+ */
+export interface LoadActiveWorkspaceDetailsTrigger {
+  (): AnyActionType;
+}
+
+/**
+ * LoadUsersOfActiveWorkspaceTrigger
+ */
+export interface LoadUsersOfActiveWorkspaceTrigger {
+  (data: {
+    workspace: WorkspaceType;
+    payload?: {
+      q: string;
+      active?: boolean;
+      firstResult?: number;
+      maxResults?: number;
+    };
+    success?: (
+      students: WorkspaceStudentListType | WorkspaceStaffListType
+    ) => void;
+    fail?: () => void;
+  }): AnyActionType;
+}
+
+/**
+ * RequestActiveWorkspaceAssessmentTrigger
+ * @param data data
+ */
+export interface RequestActiveWorkspaceAssessmentTrigger {
+  (data: {
+    workspace: WorkspaceType;
+    text: string;
+    success?: () => void;
+    fail?: () => void;
+  }): AnyActionType;
+}
+
+/**
+ * CancelAssessmentAtWorkspaceTriggerType
+ */
+export interface CancelActiveWorkspaceAssessmentRequestTrigger {
+  (data: {
+    workspace: WorkspaceType;
+    success?: () => void;
+    fail?: () => void;
+  }): AnyActionType;
 }
 
 /**
@@ -646,7 +765,7 @@ const updateWorkspaceAssessmentRequest: UpdateActiveWorkspaceAssessmentRequestTr
 const loadActiveWorkspaceChatStatus: LoadActiveWorkspaceChatStatusTrigger =
   function loadWorkspaceChatStatus() {
     return async (
-      dispatch: (arg: AnyActionType) => any,
+      dispatch: (arg: AnyActionType) => Dispatch<AnyActionType>,
       getState: () => StateType
     ) => {
       try {
@@ -683,10 +802,639 @@ const loadActiveWorkspaceChatStatus: LoadActiveWorkspaceChatStatusTrigger =
     };
   };
 
+/**
+ * loadStaffMembersOfActiveWorkspace
+ * @param data data
+ */
+const loadStaffMembersOfActiveWorkspace: LoadUsersOfActiveWorkspaceTrigger =
+  function loadStaffMembersOfActiveWorkspace(data) {
+    return async (
+      dispatch: (arg: AnyActionType) => Dispatch<AnyActionType>,
+      getState: () => StateType
+    ) => {
+      try {
+        const staffMembers = <WorkspaceStaffListType>await promisify(
+          mApi().user.staffMembers.read({
+            workspaceEntityId: data.workspace.id,
+            properties:
+              "profile-phone,profile-appointmentCalendar,profile-extraInfo,profile-whatsapp,profile-vacation-start,profile-vacation-end",
+            firstResult: data.payload ? data.payload.firstResult : 0,
+            maxResults: data.payload ? data.payload.maxResults : 10,
+          }),
+          "callback"
+        )();
+
+        const update: WorkspaceUpdateType = {
+          staffMembers,
+        };
+
+        dispatch({
+          type: "ACTIVE_WORKSPACE_UPDATE",
+          payload: {
+            original: data.workspace,
+            update,
+          },
+        });
+      } catch (err) {
+        if (!(err instanceof MApiError)) {
+          throw err;
+        }
+        dispatch(
+          displayNotification(
+            getState().i18n.text.get(
+              "plugin.workspace.management.notification.failedToLoadTeachers"
+            ),
+            "error"
+          )
+        );
+      }
+    };
+  };
+
+/**
+ * updateWorkspaceProducersForCurrentWorkspace
+ * @param data data
+ */
+const updateActiveWorkspaceProducers: UpdateActiveWorkspaceProducersTrigger =
+  function updateWorkspaceProducersForActiveWorkspace(data) {
+    return async (
+      dispatch: (arg: AnyActionType) => Dispatch<AnyActionType>,
+      getState: () => StateType
+    ) => {
+      try {
+        const state: StateType = getState();
+        const existingProducers = state.activeWorkspace.workspaceData.producers;
+
+        const workspaceProducersToAdd =
+          existingProducers.length == 0
+            ? data.appliedProducers
+            : data.appliedProducers.filter((producer) => {
+                if (!producer.id) {
+                  return producer;
+                }
+              });
+
+        const workspaceProducersToDelete = existingProducers.filter(
+          (producer) => {
+            if (producer.id) {
+              return !data.appliedProducers.find(
+                (keepProducer) => keepProducer.id === producer.id
+              );
+            }
+          }
+        );
+
+        await Promise.all(
+          workspaceProducersToAdd
+            .map((p) =>
+              promisify(
+                mApi().workspace.workspaces.materialProducers.create(
+                  state.activeWorkspace.workspaceData.id,
+                  p
+                ),
+                "callback"
+              )()
+            )
+            .concat(
+              workspaceProducersToDelete.map((p) =>
+                promisify(
+                  mApi().workspace.workspaces.materialProducers.del(
+                    state.activeWorkspace.workspaceData.id,
+                    p.id
+                  ),
+                  "callback"
+                )()
+              )
+            )
+        );
+
+        // For some reason the results of the request don't give the new workspace producers
+        // it's a mess but whatever
+        const newActualWorkspaceProducers: Array<WorkspaceProducerType> = <
+          Array<WorkspaceProducerType>
+        >await promisify(
+          mApi()
+            .workspace.workspaces.materialProducers.cacheClear()
+            .read(state.activeWorkspace.workspaceData.id),
+          "callback"
+        )();
+
+        const currentWorkspace: WorkspaceType =
+          getState().activeWorkspace.workspaceData;
+
+        dispatch({
+          type: "ACTIVE_WORKSPACE_UPDATE",
+          payload: {
+            original: currentWorkspace,
+            update: {
+              producers: newActualWorkspaceProducers,
+            },
+          },
+        });
+
+        data.success && data.success();
+      } catch (err) {
+        if (!(err instanceof MApiError)) {
+          throw err;
+        }
+        dispatch(
+          displayNotification(
+            getState().i18n.text.get(
+              "plugin.workspace.management.notification.failedToUpdateProducers"
+            ),
+            "error"
+          )
+        );
+        data.fail && data.fail();
+      }
+    };
+  };
+
+/**
+ * updateCurrentWorkspaceImagesB64
+ * @param data data
+ */
+const updateActiveWorkspaceImagesB64: UpdateActiveWorkspaceImagesB64Trigger =
+  function updateCurrentWorkspaceImagesB64(data) {
+    return async (
+      dispatch: (arg: AnyActionType) => Dispatch<AnyActionType>,
+      getState: () => StateType
+    ) => {
+      try {
+        const currentWorkspace: WorkspaceType =
+          getState().activeWorkspace.workspaceData;
+        const mimeTypeRegex = /data:([a-zA-Z0-9]+\/[a-zA-Z0-9-.+]+).*,.*/;
+        const mimeTypeOriginal =
+          data.originalB64 && data.originalB64.match(mimeTypeRegex)[1];
+        const mimeTypeCropped =
+          data.croppedB64 && data.croppedB64.match(mimeTypeRegex)[1];
+
+        if (data.delete) {
+          await promisify(
+            mApi().workspace.workspaces.workspacefile.del(
+              currentWorkspace.id,
+              "workspace-frontpage-image-cropped"
+            ),
+            "callback"
+          )();
+        } else if (data.croppedB64) {
+          await promisify(
+            mApi().workspace.workspaces.workspacefile.create(
+              currentWorkspace.id,
+              {
+                fileIdentifier: "workspace-frontpage-image-cropped",
+                contentType: mimeTypeCropped,
+                base64Data: data.croppedB64,
+              }
+            ),
+            "callback"
+          )();
+        }
+
+        if (data.delete) {
+          await promisify(
+            mApi().workspace.workspaces.workspacefile.del(
+              currentWorkspace.id,
+              "workspace-frontpage-image-original"
+            ),
+            "callback"
+          )();
+        } else if (data.originalB64) {
+          await promisify(
+            mApi().workspace.workspaces.workspacefile.create(
+              currentWorkspace.id,
+              {
+                fileIdentifier: "workspace-frontpage-image-original",
+                contentType: mimeTypeOriginal,
+                base64Data: data.originalB64,
+              }
+            ),
+            "callback"
+          )();
+        }
+
+        data.success && data.success();
+      } catch (err) {
+        if (!(err instanceof MApiError)) {
+          throw err;
+        }
+
+        dispatch(
+          displayNotification(
+            getState().i18n.text.get(
+              "plugin.workspace.management.notification.failedToUpdateImage"
+            ),
+            "error"
+          )
+        );
+
+        data.fail && data.fail();
+      }
+    };
+  };
+
+/**
+ * updateActiveWorkspaceDetails
+ * @param data data
+ */
+const updateActiveWorkspaceDetails: UpdateActiveWorkspaceDetailsTrigger =
+  function updateWorkspaceDetailsForCurrentWorkspace(data) {
+    return async (
+      dispatch: (arg: AnyActionType) => Dispatch<AnyActionType>,
+      getState: () => StateType
+    ) => {
+      try {
+        const state: StateType = getState();
+
+        await promisify(
+          mApi().workspace.workspaces.details.update(
+            state.activeWorkspace.workspaceData.id,
+            data.newDetails
+          ),
+          "callback"
+        )();
+
+        const currentWorkspace: WorkspaceType =
+          getState().activeWorkspace.workspaceData;
+
+        dispatch({
+          type: "UPDATE_WORKSPACE",
+          payload: {
+            original: currentWorkspace,
+            update: {
+              details: data.newDetails,
+            },
+          },
+        });
+
+        data.success && data.success();
+      } catch (err) {
+        if (!(err instanceof MApiError)) {
+          throw err;
+        }
+        dispatch(
+          displayNotification(
+            getState().i18n.text.get(
+              "plugin.workspace.management.notification.failedToUpdateDetails"
+            ),
+            "error"
+          )
+        );
+        data.fail && data.fail();
+      }
+    };
+  };
+
+/**
+ * loadCurrentWorkspaceUserGroupPermissions
+ */
+const loadActiveWorkspaceUserGroupPermissions: LoadActiveWorkspaceUserGroupPermissionsTrigger =
+  function loadCurrentWorkspaceUserGroupPermissions() {
+    return async (
+      dispatch: (arg: AnyActionType) => Dispatch<AnyActionType>,
+      getState: () => StateType
+    ) => {
+      try {
+        const currentWorkspace: WorkspaceType =
+          getState().activeWorkspace.workspaceData;
+        const permissions: WorkspacePermissionsType[] = <
+          WorkspacePermissionsType[]
+        >await promisify(
+          mApi().workspace.workspaces.signupGroups.read(
+            getState().activeWorkspace.workspaceData.id
+          ),
+          "callback"
+        )();
+
+        dispatch({
+          type: "ACTIVE_WORKSPACE_UPDATE",
+          payload: {
+            original: currentWorkspace,
+            update: {
+              permissions,
+            },
+          },
+        });
+      } catch (err) {
+        if (!(err instanceof MApiError)) {
+          throw err;
+        }
+        dispatch(
+          displayNotification(
+            getState().i18n.text.get(
+              "plugin.workspace.management.notification.failedToLoadWorkspacePermissions"
+            ),
+            "error"
+          )
+        );
+      }
+    };
+  };
+
+/**
+ * loadWorkspaceDetailsInCurrentWorkspace
+ */
+const loadActiveWorkspaceDetails: LoadActiveWorkspaceDetailsTrigger =
+  function loadWorkspaceDetailsInCurrentWorkspace() {
+    return async (
+      dispatch: (arg: AnyActionType) => Dispatch<AnyActionType>,
+      getState: () => StateType
+    ) => {
+      try {
+        const details: WorkspaceDetailsType = <WorkspaceDetailsType>(
+          await promisify(
+            mApi().workspace.workspaces.details.read(
+              getState().activeWorkspace.workspaceData.id
+            ),
+            "callback"
+          )()
+        );
+        const currentWorkspace: WorkspaceType =
+          getState().activeWorkspace.workspaceData;
+
+        dispatch({
+          type: "ACTIVE_WORKSPACE_UPDATE",
+          payload: {
+            original: currentWorkspace,
+            update: {
+              details,
+            },
+          },
+        });
+      } catch (err) {
+        if (!(err instanceof MApiError)) {
+          throw err;
+        }
+        dispatch(
+          displayNotification(
+            getState().i18n.text.get(
+              "plugin.workspace.management.notification.failedToLoadDetails"
+            ),
+            "error"
+          )
+        );
+      }
+    };
+  };
+
+/**
+ * loadStudentsOfWorkspace
+ * @param data data
+ */
+const loadStudentsOfActiveWorkspace: LoadUsersOfActiveWorkspaceTrigger =
+  function loadStudentsOfWorkspace(data) {
+    return async (
+      dispatch: (arg: AnyActionType) => Dispatch<AnyActionType>,
+      getState: () => StateType
+    ) => {
+      try {
+        const payload: workspaceStudentsQueryDataType = {
+          q: data.payload && data.payload.q ? data.payload.q : "",
+        };
+
+        if (data.payload && data.payload.firstResult >= 0) {
+          payload.firstResult = data.payload.firstResult;
+        }
+
+        if (data.payload && data.payload.maxResults) {
+          payload.maxResults = data.payload.maxResults;
+        }
+
+        if (data.payload && data.payload.active !== undefined) {
+          payload.active = data.payload.active;
+        }
+
+        const students = <WorkspaceStudentListType>(
+          await promisify(
+            mApi().workspace.workspaces.students.read(
+              data.workspace.id,
+              payload
+            ),
+            "callback"
+          )()
+        );
+
+        let update: WorkspaceUpdateType = {
+          students,
+        };
+
+        if (
+          data.payload &&
+          data.payload.active !== undefined &&
+          !data.payload.active
+        ) {
+          update = {
+            inactiveStudents: students,
+          };
+        }
+
+        dispatch({
+          type: "ACTIVE_WORKSPACE_UPDATE",
+          payload: {
+            original: data.workspace,
+            update,
+          },
+        });
+      } catch (err) {
+        if (!(err instanceof MApiError)) {
+          throw err;
+        }
+        dispatch(
+          displayNotification(
+            getState().i18n.text.get(
+              "plugin.workspace.management.notification.failedToLoadStudents"
+            ),
+            "error"
+          )
+        );
+      }
+    };
+  };
+
+/**
+ * requestAssessmentAtWorkspace
+ * @param data data
+ */
+const requestActiveWorkspaceAssesmentTrigger: RequestActiveWorkspaceAssessmentTrigger =
+  function requestAssessmentAtWorkspace(data) {
+    return async (
+      dispatch: (arg: AnyActionType) => Dispatch<AnyActionType>,
+      getState: () => StateType
+    ) => {
+      const state = getState();
+
+      try {
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        const assessmentRequest: WorkspaceAssessmentRequestType = <
+          WorkspaceAssessmentRequestType
+        >await promisify(
+          mApi().assessmentrequest.workspace.assessmentRequests.create(
+            data.workspace.id,
+            {
+              requestText: data.text,
+            }
+          ),
+          "callback"
+        )();
+
+        /**
+         * First finding current "assessmentState state" and depending what state it is assign new state
+         * Will be changed when module specific evaluation assessment request functionality is implemented
+         */
+        let newAssessmentState =
+          data.workspace.activity.assessmentState[0].state;
+
+        if (newAssessmentState === "unassessed") {
+          newAssessmentState = "pending";
+        } else if (newAssessmentState == "pass") {
+          newAssessmentState = "pending_pass";
+        } else if (newAssessmentState == "fail") {
+          newAssessmentState = "pending_fail";
+        } else {
+          newAssessmentState = "pending";
+        }
+
+        /**
+         * Must be done for now. To update activity when assessmentRequest is being made.
+         * In future changing state locally is better options one combination workspace module specific
+         * request are implemented
+         */
+        dispatch(updateWorkspaceActivity({}));
+
+        /**
+         * Same here
+         */
+        dispatch(updateWorkspaceAssessmentRequest({}));
+
+        dispatch(
+          actions.displayNotification(
+            state.i18n.text.get(
+              "plugin.workspace.evaluation.requestEvaluation.notificationText"
+            ),
+            "success"
+          )
+        );
+        data.success && data.success();
+      } catch (err) {
+        if (!(err instanceof MApiError)) {
+          throw err;
+        }
+        dispatch(
+          actions.displayNotification(
+            state.i18n.text.get(
+              "plugin.workspace.errormessage.requestAssessmentFail"
+            ),
+            "error"
+          )
+        );
+        data.fail && data.fail();
+      }
+    };
+  };
+
+/**
+ * cancelAssessmentAtWorkspace
+ * @param data data
+ */
+const cancelActiveWorkspaceAssessmentRequest: CancelActiveWorkspaceAssessmentRequestTrigger =
+  function cancelAssessmentAtWorkspace(data) {
+    return async (
+      dispatch: (arg: AnyActionType) => Dispatch<AnyActionType>,
+      getState: () => StateType
+    ) => {
+      const state = getState();
+
+      try {
+        const assessmentRequest: WorkspaceAssessmentRequestType =
+          data.workspace.assessmentRequests[
+            data.workspace.assessmentRequests.length - 1
+          ];
+        if (!assessmentRequest) {
+          dispatch(
+            actions.displayNotification(
+              state.i18n.text.get(
+                "plugin.workspace.errormessage.cancelAssessmentFail"
+              ),
+              "error"
+            )
+          );
+          data.fail && data.fail();
+          return;
+        }
+        await promisify(
+          mApi().assessmentrequest.workspace.assessmentRequests.del(
+            data.workspace.id,
+            assessmentRequest.id
+          ),
+          "callback"
+        )();
+
+        /**
+         * First finding current "assessmentState state" and depending what state it is assign new state
+         * Will be changed when module specific evaluation assessment request functionality is implemented
+         */
+        let newAssessmentState =
+          data.workspace.activity.assessmentState[0].state;
+
+        if (newAssessmentState == "pending") {
+          newAssessmentState = "unassessed";
+        } else if (newAssessmentState == "pending_pass") {
+          newAssessmentState = "pass";
+        } else if (newAssessmentState == "pending_fail") {
+          newAssessmentState = "fail";
+        }
+
+        /**
+         * Must be done for now. To update activity when assessmentRequest is being made.
+         * In future changing state locally is better options one combination workspace module specific
+         * request are implemented
+         */
+        dispatch(updateWorkspaceActivity({}));
+
+        /**
+         * Same here
+         */
+        dispatch(updateWorkspaceAssessmentRequest({}));
+
+        dispatch(
+          actions.displayNotification(
+            state.i18n.text.get(
+              "plugin.workspace.evaluation.cancelEvaluation.notificationText"
+            ),
+            "success"
+          )
+        );
+        data.success && data.success();
+      } catch (err) {
+        if (!(err instanceof MApiError)) {
+          throw err;
+        }
+        dispatch(
+          actions.displayNotification(
+            state.i18n.text.get(
+              "plugin.workspace.errormessage.cancelAssessmentFail"
+            ),
+            "error"
+          )
+        );
+        data.fail && data.fail();
+      }
+    };
+  };
+
 export {
   setActiveWorkspace,
   updateActiveWorkspace,
   updateWorkspaceActivity,
   updateWorkspaceAssessmentRequest,
+  updateActiveWorkspaceProducers,
+  updateActiveWorkspaceImagesB64,
+  updateActiveWorkspaceDetails,
   loadActiveWorkspaceChatStatus,
+  loadStaffMembersOfActiveWorkspace,
+  loadActiveWorkspaceUserGroupPermissions,
+  loadActiveWorkspaceDetails,
+  loadStudentsOfActiveWorkspace,
+  requestActiveWorkspaceAssesmentTrigger,
+  cancelActiveWorkspaceAssessmentRequest,
 };
