+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/actions/workspaces/index.ts
@@ -31,7 +31,6 @@ import {
   MaterialContentNodeListType,
   MaterialCompositeRepliesListType,
   MaterialCompositeRepliesStateType,
-  WorkspaceJournalsType,
   WorkspaceDetailsType,
   WorkspaceTypeType,
   WorkspaceProducerType,
@@ -463,38 +462,6 @@ export interface SetAvailableCurriculumsTriggerType {
 }
 
 /**
- * UpdateCurrentWorkspaceImagesB64TriggerType
- */
-export interface UpdateCurrentWorkspaceImagesB64TriggerType {
-  (data?: {
-    delete?: boolean;
-    originalB64?: string;
-    croppedB64?: string;
-    success?: () => any;
-    fail?: () => any;
-  }): AnyActionType;
-}
-
-/**
- * LoadCurrentWorkspaceUserGroupPermissionsTriggerType
- */
-export interface LoadCurrentWorkspaceUserGroupPermissionsTriggerType {
-  (): AnyActionType;
-}
-
-/**
- * UpdateCurrentWorkspaceUserGroupPermissionTriggerType
- */
-export interface UpdateCurrentWorkspaceUserGroupPermissionTriggerType {
-  (data?: {
-    original: WorkspacePermissionsType;
-    update: WorkspacePermissionsType;
-    success?: () => any;
-    fail?: () => any;
-  }): AnyActionType;
-}
-
-/**
  * SetWorkspaceMaterialEditorStateTriggerType
  */
 export interface SetWorkspaceMaterialEditorStateTriggerType {
@@ -676,208 +643,6 @@ const setCurrentOrganizationWorkspace: SetCurrentWorkspaceTriggerType =
   };
 
 /**
- * setCurrentWorkspace
- * @param data data
- */
-const setCurrentWorkspace: SetCurrentWorkspaceTriggerType =
-  function setCurrentWorkspace(data) {
-    return async (
-      dispatch: (arg: AnyActionType) => any,
-      getState: () => StateType
-    ) => {
-      const state = getState();
-
-      const current: WorkspaceType = state.activeWorkspace.workspaceData;
-      if (
-        current &&
-        current.id === data.workspaceId &&
-        !data.refreshActivity &&
-        !data.loadDetails
-      ) {
-        data.success && data.success(current);
-        return;
-      }
-
-      try {
-        let workspace: WorkspaceType =
-          state.workspaces.userWorkspaces.find(
-            (w) => w.id === data.workspaceId
-          ) ||
-          state.workspaces.availableWorkspaces.find(
-            (w) => w.id === data.workspaceId
-          );
-        if (current && current.id === data.workspaceId) {
-          //if I just make it be current it will be buggy
-          workspace = { ...current };
-        }
-
-        /* let assesments: WorkspaceStudentAssessmentsType; */
-        let assessmentRequests: Array<WorkspaceAssessmentRequestType>;
-        let activity: WorkspaceActivityType;
-        let additionalInfo: WorkspaceAdditionalInfoType;
-        let contentDescription: MaterialContentNodeType;
-        let producers: Array<WorkspaceProducerType>;
-        let isCourseMember: boolean;
-        let journals: WorkspaceJournalsType;
-        let details: WorkspaceDetailsType;
-        let chatStatus: WorkspaceChatStatusType;
-        const status = state.status;
-
-        [
-          workspace,
-          assessmentRequests,
-          activity,
-          additionalInfo,
-          contentDescription,
-          producers,
-          isCourseMember,
-          journals,
-          details,
-          chatStatus,
-        ] = (await Promise.all([
-          reuseExistantValue(true, workspace, () =>
-            promisify(
-              mApi().workspace.workspaces.cacheClear().read(data.workspaceId),
-              "callback"
-            )()
-          ),
-
-          reuseExistantValue(
-            status.permissions.WORKSPACE_REQUEST_WORKSPACE_ASSESSMENT,
-            workspace && workspace.assessmentRequests,
-            () =>
-              promisify(
-                mApi()
-                  .assessmentrequest.workspace.assessmentRequests.cacheClear()
-                  .read(data.workspaceId, {
-                    studentIdentifier: state.status.userSchoolDataIdentifier,
-                  }),
-                "callback"
-              )()
-          ),
-
-          status.loggedIn
-            ? // The way refresh works is by never giving an existant value to the reuse existant value function that way it will think that there's no value
-              // And rerequest
-              reuseExistantValue(
-                true,
-                typeof data.refreshActivity !== "undefined" &&
-                  data.refreshActivity
-                  ? null
-                  : workspace && workspace.activity,
-                () =>
-                  promisify(
-                    mApi()
-                      .evaluation.workspaces.students.activity.cacheClear()
-                      .read(
-                        data.workspaceId,
-                        state.status.userSchoolDataIdentifier
-                      ),
-                    "callback"
-                  )()
-              )
-            : null,
-
-          reuseExistantValue(true, workspace && workspace.additionalInfo, () =>
-            promisify(
-              mApi()
-                .workspace.workspaces.additionalInfo.cacheClear()
-                .read(data.workspaceId),
-              "callback"
-            )()
-          ),
-
-          reuseExistantValue(
-            true,
-            workspace && workspace.contentDescription,
-            () =>
-              promisify(
-                mApi()
-                  .workspace.workspaces.description.cacheClear()
-                  .read(data.workspaceId),
-                "callback"
-              )()
-          ),
-
-          reuseExistantValue(true, workspace && workspace.producers, () =>
-            promisify(
-              mApi()
-                .workspace.workspaces.materialProducers.cacheClear()
-                .read(data.workspaceId),
-              "callback"
-            )()
-          ),
-
-          state.status.loggedIn
-            ? reuseExistantValue(
-                true,
-                workspace &&
-                  typeof workspace.isCourseMember !== "undefined" &&
-                  workspace.isCourseMember,
-                () =>
-                  promisify(
-                    mApi().workspace.workspaces.amIMember.read(
-                      data.workspaceId
-                    ),
-                    "callback"
-                  )()
-              )
-            : false,
-
-          reuseExistantValue(true, workspace && workspace.journals, () => null),
-
-          data.loadDetails || (workspace && workspace.details)
-            ? reuseExistantValue(true, workspace && workspace.details, () =>
-                promisify(
-                  mApi().workspace.workspaces.details.read(data.workspaceId),
-                  "callback"
-                )()
-              )
-            : null,
-
-          state.status.loggedIn
-            ? reuseExistantValue(true, workspace && workspace.chatStatus, () =>
-                promisify(
-                  mApi().chat.workspaceChatSettings.read(data.workspaceId),
-                  "callback"
-                )()
-              )
-            : null,
-        ])) as any;
-        workspace.assessmentRequests = assessmentRequests;
-        workspace.activity = activity;
-        workspace.additionalInfo = additionalInfo;
-        workspace.contentDescription = contentDescription;
-        workspace.producers = producers;
-        workspace.isCourseMember = isCourseMember;
-        workspace.journals = journals;
-        workspace.details = details;
-        workspace.chatStatus = chatStatus;
-
-        dispatch({
-          type: "SET_CURRENT_WORKSPACE",
-          payload: workspace,
-        });
-
-        data.success && data.success(workspace);
-      } catch (err) {
-        if (!(err instanceof MApiError)) {
-          throw err;
-        }
-        dispatch(
-          actions.displayNotification(
-            state.i18n.text.get(
-              "plugin.workspace.errormessage.workspaceLoadFailed"
-            ),
-            "error"
-          )
-        );
-        data.fail && data.fail();
-      }
-    };
-  };
-
-/**
  * setAvailableCurriculums
  * @returns Promise<void>
  */
@@ -910,59 +675,6 @@ const setAvailableCurriculums: SetAvailableCurriculumsTriggerType =
             "error"
           )
         );
-      }
-    };
-  };
-
-/**
- * UpdateCurrentWorkspaceActivityTriggerType
- */
-export interface UpdateCurrentWorkspaceActivityTriggerType {
-  (data: { success?: () => any; fail?: () => any }): AnyActionType;
-}
-
-/**
- * This is a thunk action creator that returns a thunk function
- *
- * @param data data
- * @returns a thunk function
- */
-const updateCurrentWorkspaceActivity: UpdateCurrentWorkspaceActivityTriggerType =
-  function updateCurrentWorkspaceActivity(data) {
-    return async (
-      dispatch: (arg: AnyActionType) => any,
-      getState: () => StateType
-    ) => {
-      const state = getState();
-
-      if (state.status.loggedIn) {
-        try {
-          const activity = <WorkspaceActivityType>(
-            await promisify(
-              mApi()
-                .evaluation.workspaces.students.activity.cacheClear()
-                .read(
-                  state.activeWorkspace.workspaceData.id,
-                  state.status.userSchoolDataIdentifier
-                ),
-              "callback"
-            )()
-          );
-
-          dispatch({
-            type: "UPDATE_CURRENT_WORKSPACE_ACTIVITY",
-            payload: activity,
-          });
-        } catch (err) {
-          dispatch(
-            actions.displayNotification(
-              state.i18n.text.get(
-                "plugin.workspace.errormessage.workspaceActivityLoadFailed"
-              ),
-              "error"
-            )
-          );
-        }
       }
     };
   };
@@ -1015,215 +727,6 @@ const updateCurrentWorkspaceAssessmentRequest: UpdateCurrentWorkspaceAssessmentR
             )
           );
         }
-      }
-    };
-  };
-
-/**
- * RequestAssessmentAtWorkspaceTriggerType
- */
-export interface RequestAssessmentAtWorkspaceTriggerType {
-  (data: {
-    workspace: WorkspaceType;
-    text: string;
-    success?: () => any;
-    fail?: () => any;
-  }): AnyActionType;
-}
-
-/**
- * requestAssessmentAtWorkspace
- * @param data data
- */
-/**
- * requestAssessmentAtWorkspace
- * @param data data
- */
-export interface RequestAssessmentAtWorkspaceTriggerType {
-  (data: {
-    workspace: WorkspaceType;
-    text: string;
-    success?: () => any;
-    fail?: () => any;
-  }): AnyActionType;
-}
-/**
- * requestAssessmentAtWorkspace
- * @param data data
- */
-const requestAssessmentAtWorkspace: RequestAssessmentAtWorkspaceTriggerType =
-  function requestAssessmentAtWorkspace(data) {
-    return async (
-      dispatch: (arg: AnyActionType) => any,
-      getState: () => StateType
-    ) => {
-      const state = getState();
-
-      try {
-        const assessmentRequest: WorkspaceAssessmentRequestType = <
-          WorkspaceAssessmentRequestType
-        >await promisify(
-          mApi().assessmentrequest.workspace.assessmentRequests.create(
-            data.workspace.id,
-            {
-              requestText: data.text,
-            }
-          ),
-          "callback"
-        )();
-
-        /**
-         * First finding current "assessmentState state" and depending what state it is assign new state
-         * Will be changed when module specific evaluation assessment request functionality is implemented
-         */
-        let newAssessmentState =
-          data.workspace.activity.assessmentState[0].state;
-
-        if (newAssessmentState === "unassessed") {
-          newAssessmentState = "pending";
-        } else if (newAssessmentState == "pass") {
-          newAssessmentState = "pending_pass";
-        } else if (newAssessmentState == "fail") {
-          newAssessmentState = "pending_fail";
-        } else {
-          newAssessmentState = "pending";
-        }
-
-        /**
-         * Must be done for now. To update activity when assessmentRequest is being made.
-         * In future changing state locally is better options one combination workspace module specific
-         * request are implemented
-         */
-        dispatch(updateCurrentWorkspaceActivity({}));
-
-        /**
-         * Same here
-         */
-        dispatch(updateCurrentWorkspaceAssessmentRequest({}));
-
-        dispatch(
-          actions.displayNotification(
-            state.i18n.text.get(
-              "plugin.workspace.evaluation.requestEvaluation.notificationText"
-            ),
-            "success"
-          )
-        );
-        data.success && data.success();
-      } catch (err) {
-        if (!(err instanceof MApiError)) {
-          throw err;
-        }
-        dispatch(
-          actions.displayNotification(
-            state.i18n.text.get(
-              "plugin.workspace.errormessage.requestAssessmentFail"
-            ),
-            "error"
-          )
-        );
-        data.fail && data.fail();
-      }
-    };
-  };
-
-/**
- * CancelAssessmentAtWorkspaceTriggerType
- */
-export interface CancelAssessmentAtWorkspaceTriggerType {
-  (data: {
-    workspace: WorkspaceType;
-    success?: () => any;
-    fail?: () => any;
-  }): AnyActionType;
-}
-
-/**
- * cancelAssessmentAtWorkspace
- * @param data data
- */
-const cancelAssessmentAtWorkspace: CancelAssessmentAtWorkspaceTriggerType =
-  function cancelAssessmentAtWorkspace(data) {
-    return async (
-      dispatch: (arg: AnyActionType) => any,
-      getState: () => StateType
-    ) => {
-      const state = getState();
-
-      try {
-        const assessmentRequest: WorkspaceAssessmentRequestType =
-          data.workspace.assessmentRequests[
-            data.workspace.assessmentRequests.length - 1
-          ];
-        if (!assessmentRequest) {
-          dispatch(
-            actions.displayNotification(
-              state.i18n.text.get(
-                "plugin.workspace.errormessage.cancelAssessmentFail"
-              ),
-              "error"
-            )
-          );
-          data.fail && data.fail();
-          return;
-        }
-        await promisify(
-          mApi().assessmentrequest.workspace.assessmentRequests.del(
-            data.workspace.id,
-            assessmentRequest.id
-          ),
-          "callback"
-        )();
-
-        /**
-         * First finding current "assessmentState state" and depending what state it is assign new state
-         * Will be changed when module specific evaluation assessment request functionality is implemented
-         */
-        let newAssessmentState =
-          data.workspace.activity.assessmentState[0].state;
-
-        if (newAssessmentState == "pending") {
-          newAssessmentState = "unassessed";
-        } else if (newAssessmentState == "pending_pass") {
-          newAssessmentState = "pass";
-        } else if (newAssessmentState == "pending_fail") {
-          newAssessmentState = "fail";
-        }
-
-        /**
-         * Must be done for now. To update activity when assessmentRequest is being made.
-         * In future changing state locally is better options one combination workspace module specific
-         * request are implemented
-         */
-        dispatch(updateCurrentWorkspaceActivity({}));
-
-        /**
-         * Same here
-         */
-        dispatch(updateCurrentWorkspaceAssessmentRequest({}));
-
-        dispatch(
-          actions.displayNotification(
-            state.i18n.text.get(
-              "plugin.workspace.evaluation.cancelEvaluation.notificationText"
-            ),
-            "success"
-          )
-        );
-        data.success && data.success();
-      } catch (err) {
-        if (!(err instanceof MApiError)) {
-          throw err;
-        }
-        dispatch(
-          actions.displayNotification(
-            state.i18n.text.get(
-              "plugin.workspace.errormessage.cancelAssessmentFail"
-            ),
-            "error"
-          )
-        );
-        data.fail && data.fail();
       }
     };
   };
@@ -2058,55 +1561,6 @@ const loadCurrentOrganizationWorkspaceStaff: LoadUsersOfWorkspaceTriggerType =
   };
 
 /**
- * loadStaffMembersOfWorkspace
- * @param data data
- */
-const loadStaffMembersOfWorkspace: LoadUsersOfWorkspaceTriggerType =
-  function loadStaffMembersOfWorkspace(data) {
-    return async (
-      dispatch: (arg: AnyActionType) => any,
-      getState: () => StateType
-    ) => {
-      try {
-        const staffMembers = <WorkspaceStaffListType>await promisify(
-          mApi().user.staffMembers.read({
-            workspaceEntityId: data.workspace.id,
-            properties:
-              "profile-phone,profile-appointmentCalendar,profile-extraInfo,profile-whatsapp,profile-vacation-start,profile-vacation-end",
-            firstResult: data.payload ? data.payload.firstResult : 0,
-            maxResults: data.payload ? data.payload.maxResults : 10,
-          }),
-          "callback"
-        )();
-
-        const update: WorkspaceUpdateType = {
-          staffMembers,
-        };
-
-        dispatch({
-          type: "UPDATE_WORKSPACE",
-          payload: {
-            original: data.workspace,
-            update,
-          },
-        });
-      } catch (err) {
-        if (!(err instanceof MApiError)) {
-          throw err;
-        }
-        dispatch(
-          displayNotification(
-            getState().i18n.text.get(
-              "plugin.workspace.management.notification.failedToLoadTeachers"
-            ),
-            "error"
-          )
-        );
-      }
-    };
-  };
-
-/**
  * loadCurrentOrganizationWorkspaceStudents
  * @param data data
  */
@@ -2162,80 +1616,6 @@ const loadCurrentOrganizationWorkspaceStudents: LoadUsersOfWorkspaceTriggerType 
           type: "UPDATE_ORGANIZATION_SELECTED_WORKSPACE",
           payload: { studentsSelect: { state: "ERROR" } },
         });
-      }
-    };
-  };
-
-/**
- * loadStudentsOfWorkspace
- * @param data data
- */
-const loadStudentsOfWorkspace: LoadUsersOfWorkspaceTriggerType =
-  function loadStudentsOfWorkspace(data) {
-    return async (
-      dispatch: (arg: AnyActionType) => any,
-      getState: () => StateType
-    ) => {
-      try {
-        const payload: workspaceStudentsQueryDataType = {
-          q: data.payload && data.payload.q ? data.payload.q : "",
-        };
-
-        if (data.payload && data.payload.firstResult >= 0) {
-          payload.firstResult = data.payload.firstResult;
-        }
-
-        if (data.payload && data.payload.maxResults) {
-          payload.maxResults = data.payload.maxResults;
-        }
-
-        if (data.payload && data.payload.active !== undefined) {
-          payload.active = data.payload.active;
-        }
-
-        const students = <WorkspaceStudentListType>(
-          await promisify(
-            mApi().workspace.workspaces.students.read(
-              data.workspace.id,
-              payload
-            ),
-            "callback"
-          )()
-        );
-
-        let update: WorkspaceUpdateType = {
-          students,
-        };
-
-        if (
-          data.payload &&
-          data.payload.active !== undefined &&
-          !data.payload.active
-        ) {
-          update = {
-            inactiveStudents: students,
-          };
-        }
-
-        dispatch({
-          type: "UPDATE_WORKSPACE",
-          payload: {
-            original: data.workspace,
-            update,
-          },
-        });
-      } catch (err) {
-        if (!(err instanceof MApiError)) {
-          throw err;
-        }
-        dispatch(
-          displayNotification(
-            getState().i18n.text.get(
-              "plugin.workspace.management.notification.failedToLoadStudents"
-            ),
-            "error"
-          )
-        );
       }
     };
   };
@@ -2591,32 +1971,10 @@ export interface LoadWorkspaceDetailsInCurrentWorkspaceTriggerType {
 }
 
 /**
- * UpdateWorkspaceDetailsForCurrentWorkspaceTriggerType
- */
-export interface UpdateWorkspaceDetailsForCurrentWorkspaceTriggerType {
-  (data: {
-    newDetails: WorkspaceDetailsType;
-    success: () => any;
-    fail: () => any;
-  }): AnyActionType;
-}
-
-/**
  * LoadWorkspaceChatStatusTriggerType
  */
 export interface LoadWorkspaceChatStatusTriggerType {
   (): AnyActionType;
-}
-
-/**
- * UpdateWorkspaceProducersForCurrentWorkspaceTriggerType
- */
-export interface UpdateWorkspaceProducersForCurrentWorkspaceTriggerType {
-  (data: {
-    appliedProducers: Array<WorkspaceProducerType>;
-    success: () => any;
-    fail: () => any;
-  }): AnyActionType;
 }
 
 /**
@@ -2737,203 +2095,6 @@ const loadWorkspaceChatStatus: LoadWorkspaceChatStatusTriggerType =
             "error"
           )
         );
-      }
-    };
-  };
-
-/**
- * loadWorkspaceDetailsInCurrentWorkspace
- */
-const loadWorkspaceDetailsInCurrentWorkspace: LoadWorkspaceDetailsInCurrentWorkspaceTriggerType =
-  function loadWorkspaceDetailsInCurrentWorkspace() {
-    return async (
-      dispatch: (arg: AnyActionType) => any,
-      getState: () => StateType
-    ) => {
-      try {
-        const details: WorkspaceDetailsType = <WorkspaceDetailsType>(
-          await promisify(
-            mApi().workspace.workspaces.details.read(
-              getState().activeWorkspace.workspaceData.id
-            ),
-            "callback"
-          )()
-        );
-        const currentWorkspace: WorkspaceType =
-          getState().activeWorkspace.workspaceData;
-
-        dispatch({
-          type: "UPDATE_WORKSPACE",
-          payload: {
-            original: currentWorkspace,
-            update: {
-              details,
-            },
-          },
-        });
-      } catch (err) {
-        if (!(err instanceof MApiError)) {
-          throw err;
-        }
-        dispatch(
-          displayNotification(
-            getState().i18n.text.get(
-              "plugin.workspace.management.notification.failedToLoadDetails"
-            ),
-            "error"
-          )
-        );
-      }
-    };
-  };
-
-/**
- * updateWorkspaceDetailsForCurrentWorkspace
- * @param data data
- */
-const updateWorkspaceDetailsForCurrentWorkspace: UpdateWorkspaceDetailsForCurrentWorkspaceTriggerType =
-  function updateWorkspaceDetailsForCurrentWorkspace(data) {
-    return async (
-      dispatch: (arg: AnyActionType) => any,
-      getState: () => StateType
-    ) => {
-      try {
-        const state: StateType = getState();
-
-        await promisify(
-          mApi().workspace.workspaces.details.update(
-            state.activeWorkspace.workspaceData.id,
-            data.newDetails
-          ),
-          "callback"
-        )();
-
-        const currentWorkspace: WorkspaceType =
-          getState().activeWorkspace.workspaceData;
-
-        dispatch({
-          type: "UPDATE_WORKSPACE",
-          payload: {
-            original: currentWorkspace,
-            update: {
-              details: data.newDetails,
-            },
-          },
-        });
-
-        data.success && data.success();
-      } catch (err) {
-        if (!(err instanceof MApiError)) {
-          throw err;
-        }
-        dispatch(
-          displayNotification(
-            getState().i18n.text.get(
-              "plugin.workspace.management.notification.failedToUpdateDetails"
-            ),
-            "error"
-          )
-        );
-        data.fail && data.fail();
-      }
-    };
-  };
-
-/**
- * updateWorkspaceProducersForCurrentWorkspace
- * @param data data
- */
-const updateWorkspaceProducersForCurrentWorkspace: UpdateWorkspaceProducersForCurrentWorkspaceTriggerType =
-  function updateWorkspaceProducersForCurrentWorkspace(data) {
-    return async (
-      dispatch: (arg: AnyActionType) => any,
-      getState: () => StateType
-    ) => {
-      try {
-        const state: StateType = getState();
-        const existingProducers = state.activeWorkspace.workspaceData.producers;
-
-        const workspaceProducersToAdd =
-          existingProducers.length == 0
-            ? data.appliedProducers
-            : data.appliedProducers.filter((producer) => {
-                if (!producer.id) {
-                  return producer;
-                }
-              });
-
-        const workspaceProducersToDelete = existingProducers.filter(
-          (producer) => {
-            if (producer.id) {
-              return !data.appliedProducers.find(
-                (keepProducer) => keepProducer.id === producer.id
-              );
-            }
-          }
-        );
-
-        await Promise.all(
-          workspaceProducersToAdd
-            .map((p) =>
-              promisify(
-                mApi().workspace.workspaces.materialProducers.create(
-                  state.activeWorkspace.workspaceData.id,
-                  p
-                ),
-                "callback"
-              )()
-            )
-            .concat(
-              workspaceProducersToDelete.map((p) =>
-                promisify(
-                  mApi().workspace.workspaces.materialProducers.del(
-                    state.activeWorkspace.workspaceData.id,
-                    p.id
-                  ),
-                  "callback"
-                )()
-              )
-            )
-        );
-
-        // For some reason the results of the request don't give the new workspace producers
-        // it's a mess but whatever
-        const newActualWorkspaceProducers: Array<WorkspaceProducerType> = <
-          Array<WorkspaceProducerType>
-        >await promisify(
-          mApi()
-            .workspace.workspaces.materialProducers.cacheClear()
-            .read(state.activeWorkspace.workspaceData.id),
-          "callback"
-        )();
-
-        const currentWorkspace: WorkspaceType =
-          getState().activeWorkspace.workspaceData;
-
-        dispatch({
-          type: "UPDATE_WORKSPACE",
-          payload: {
-            original: currentWorkspace,
-            update: {
-              producers: newActualWorkspaceProducers,
-            },
-          },
-        });
-
-        data.success && data.success();
-      } catch (err) {
-        if (!(err instanceof MApiError)) {
-          throw err;
-        }
-        dispatch(
-          displayNotification(
-            getState().i18n.text.get(
-              "plugin.workspace.management.notification.failedToUpdateProducers"
-            ),
-            "error"
-          )
-        );
-        data.fail && data.fail();
       }
     };
   };
@@ -3236,135 +2397,6 @@ const copyCurrentWorkspace: CopyCurrentWorkspaceTriggerType =
         );
 
         data.fail && data.fail();
-      }
-    };
-  };
-
-/**
- * updateCurrentWorkspaceImagesB64
- * @param data data
- */
-const updateCurrentWorkspaceImagesB64: UpdateCurrentWorkspaceImagesB64TriggerType =
-  function updateCurrentWorkspaceImagesB64(data) {
-    return async (
-      dispatch: (arg: AnyActionType) => any,
-      getState: () => StateType
-    ) => {
-      try {
-        const currentWorkspace: WorkspaceType =
-          getState().activeWorkspace.workspaceData;
-        const mimeTypeRegex = /data:([a-zA-Z0-9]+\/[a-zA-Z0-9-.+]+).*,.*/;
-        const mimeTypeOriginal =
-          data.originalB64 && data.originalB64.match(mimeTypeRegex)[1];
-        const mimeTypeCropped =
-          data.croppedB64 && data.croppedB64.match(mimeTypeRegex)[1];
-
-        if (data.delete) {
-          await promisify(
-            mApi().workspace.workspaces.workspacefile.del(
-              currentWorkspace.id,
-              "workspace-frontpage-image-cropped"
-            ),
-            "callback"
-          )();
-        } else if (data.croppedB64) {
-          await promisify(
-            mApi().workspace.workspaces.workspacefile.create(
-              currentWorkspace.id,
-              {
-                fileIdentifier: "workspace-frontpage-image-cropped",
-                contentType: mimeTypeCropped,
-                base64Data: data.croppedB64,
-              }
-            ),
-            "callback"
-          )();
-        }
-
-        if (data.delete) {
-          await promisify(
-            mApi().workspace.workspaces.workspacefile.del(
-              currentWorkspace.id,
-              "workspace-frontpage-image-original"
-            ),
-            "callback"
-          )();
-        } else if (data.originalB64) {
-          await promisify(
-            mApi().workspace.workspaces.workspacefile.create(
-              currentWorkspace.id,
-              {
-                fileIdentifier: "workspace-frontpage-image-original",
-                contentType: mimeTypeOriginal,
-                base64Data: data.originalB64,
-              }
-            ),
-            "callback"
-          )();
-        }
-
-        data.success && data.success();
-      } catch (err) {
-        if (!(err instanceof MApiError)) {
-          throw err;
-        }
-
-        dispatch(
-          displayNotification(
-            getState().i18n.text.get(
-              "plugin.workspace.management.notification.failedToUpdateImage"
-            ),
-            "error"
-          )
-        );
-
-        data.fail && data.fail();
-      }
-    };
-  };
-
-/**
- * loadCurrentWorkspaceUserGroupPermissions
- */
-const loadCurrentWorkspaceUserGroupPermissions: LoadCurrentWorkspaceUserGroupPermissionsTriggerType =
-  function loadCurrentWorkspaceUserGroupPermissions() {
-    return async (
-      dispatch: (arg: AnyActionType) => any,
-      getState: () => StateType
-    ) => {
-      try {
-        const currentWorkspace: WorkspaceType =
-          getState().activeWorkspace.workspaceData;
-        const permissions: WorkspacePermissionsType[] = <
-          WorkspacePermissionsType[]
-        >await promisify(
-          mApi().workspace.workspaces.signupGroups.read(
-            getState().activeWorkspace.workspaceData.id
-          ),
-          "callback"
-        )();
-
-        dispatch({
-          type: "UPDATE_WORKSPACE",
-          payload: {
-            original: currentWorkspace,
-            update: {
-              permissions,
-            },
-          },
-        });
-      } catch (err) {
-        if (!(err instanceof MApiError)) {
-          throw err;
-        }
-        dispatch(
-          displayNotification(
-            getState().i18n.text.get(
-              "plugin.workspace.management.notification.failedToLoadWorkspacePermissions"
-            ),
-            "error"
-          )
-        );
       }
     };
   };
@@ -4235,29 +3267,19 @@ export {
   signupIntoWorkspace,
   loadUserWorkspacesFromServer,
   loadLastWorkspaceFromServer,
-  setCurrentWorkspace,
-  requestAssessmentAtWorkspace,
-  cancelAssessmentAtWorkspace,
   updateWorkspace,
-  loadStaffMembersOfWorkspace,
   loadWorkspaceChatStatus,
   loadWholeWorkspaceMaterials,
   setCurrentWorkspaceMaterialsActiveNodeId,
   loadWorkspaceCompositeMaterialReplies,
   updateAssignmentState,
   updateLastWorkspace,
-  loadStudentsOfWorkspace,
   toggleActiveStateOfStudentOfWorkspace,
   createWorkspace,
   updateOrganizationWorkspace,
-  loadWorkspaceDetailsInCurrentWorkspace,
   loadWorkspaceTypes,
   deleteCurrentWorkspaceImage,
   copyCurrentWorkspace,
-  updateWorkspaceDetailsForCurrentWorkspace,
-  updateWorkspaceProducersForCurrentWorkspace,
-  updateCurrentWorkspaceImagesB64,
-  loadCurrentWorkspaceUserGroupPermissions,
   setWorkspaceMaterialEditorState,
   updateWorkspaceMaterialContentNode,
   deleteWorkspaceMaterialContentNode,
@@ -4269,7 +3291,6 @@ export {
   loadTemplatesFromServer,
   updateWorkspaceEditModeState,
   loadWholeWorkspaceHelp,
-  updateCurrentWorkspaceActivity,
   updateCurrentWorkspaceAssessmentRequest,
   setWholeWorkspaceHelp,
   setAvailableCurriculums,
