+++ muikku-core-plugins/src/main/resources/META-INF/resources/scripts/src/components/workspace/workspaceMaterials/materials/material.tsx
@@ -10,10 +10,6 @@ import {
   MaterialCompositeRepliesType,
   WorkspaceEditModeStateType,
 } from "~/reducers/workspaces";
-import {
-  setCurrentWorkspace,
-  SetCurrentWorkspaceTriggerType,
-} from "~/actions/workspaces";
 import { bindActionCreators } from "redux";
 import { MaterialLoaderEditorButtonSet } from "~/components/base/material-loader/editor-buttonset";
 import { MaterialLoaderTitle } from "~/components/base/material-loader/title";
@@ -45,7 +41,6 @@ interface WorkspaceMaterialProps {
   isViewRestricted: boolean;
   showEvenIfHidden: boolean;
   workspace: WorkspaceType;
-  setCurrentWorkspace: SetCurrentWorkspaceTriggerType;
   setActiveWorkspace: SetActiveWorkspaceTrigger;
 }
 
@@ -77,12 +72,6 @@ class WorkspaceMaterial extends React.Component<
     //This function is very efficient and reuses as much data as possible so
     // it won't call anything from the server other than
     //to refresh the activity and that's because we are forcing it to do so
-
-    // TODO: REMOVE
-    this.props.setCurrentWorkspace({
-      workspaceId: this.props.workspace.id,
-      refreshActivity: true,
-    });
 
     this.props.setActiveWorkspace({
       workspaceId: this.props.workspace.id,
@@ -221,10 +210,7 @@ function mapStateToProps(state: StateType) {
  * @param dispatch dispatch
  */
 function mapDispatchToProps(dispatch: Dispatch<AnyActionType>) {
-  return bindActionCreators(
-    { setCurrentWorkspace, setActiveWorkspace },
-    dispatch
-  );
+  return bindActionCreators({ setActiveWorkspace }, dispatch);
 }
 
 export default connect(mapStateToProps, mapDispatchToProps)(WorkspaceMaterial);
